AUDIT LOGS!
We have reached audit logs, at least mostly. The database migration is not fully written yet, but we're done on the rust side of things.
This commit is freaking huge because I worked on-and-off-again on it during the past month or so cause uni schedule is crazy.
Anyway, audit_logs work in the way that they are generated completely on the SQL side of things.
The rust code communicated with the SQL functions via a temporary table, into which it inserts the id of the current user (this works, since temporary tables are per-connection). Audit log entries are always attributed to the user id currently in that table.
If the table does not exist, it falls back to a global table by the same name which contains  a dummy user to which those entries will be attributed.

I've also abstracted away the permission handling which should make the actual handler functions even shorter and even allow us some fancy things like determining permissions based on the content of some patch data