Wrote macros to generate the code for the common request handlers
like the common GET, PATCH and DELETE stuff that was literally just copied into each and every file. Also made it easier to implement almost all of the missing handlers. Only thing left now is PATCH /record/[record id]/, though for that I'll first have to write the Patch impl for Record, which I do not have the time for currently