Fixed panic because I forgot to wrap the if_match in a some
Dammit type inference! So here's what happened:  The method expected an Option<IfMatch>. The value I passed it was simple the if_match variable retrieved earlier. How was it retrieved. Using req.extensions_mut().remove(). The problem is, that that methods return type is compatible with everything. What happened was that it tried to lookup Option<IfMatch> in the extensions dict, which obviously wasn't there. We put in a normal IfMatch after all. Calling unwrap resulted in the panic

To be safe I've annotated every occurence of this explicitly now.