implemented a system for making notes on records
Whenever a record is deleted, and another record is the cause for that (i.e. a record with higher progress gets approved, replacing the current approved records), all notes will automatically move over.

This included a revamp of how record modifications are handled (the transitions between record states are now well-defined) and also the addition of a fourth record state ('under consideration').

Additionally, submitters can now leave an initial note on records.