Fixed audit log entries caused by unauthenticated connection being attributed wrongly
They are supposed to be attributed to a dummy user with id 0. Instead they were attributed to whichever authenticated user last used the database connection. This happened because the connection metadata regarding audit-log-user-id was not reset when the connection was returned to the pool. Now even unauthenticated connections explicitly set the user id to 0.