fix: Move registration ratelimiter after "name taken" check
Triggering the ratelimit before the call to
`AuthenticatedUser::register` would cause failed registration attempts
(due to the user name already being taken) to count towards the hard
limit of "1 registration per day". Instead, only trigger the ratelimiter
after the call to `register`, and rely on transaction roleback (via
returning before the `commit` call) to undo the registration in case a
ratelimit is triggered.