fix: Do not mix server and user timezones in time machine
By defining `beginning_of_time` as a `DateTime<FixedOffset>`, it was
defined in terms of the server's time zone (which is whatever is going
on in western Europe these days). However, the time requested by the
user is always in their local time zone, so make sure we do not mix this
(by always using UTC on the server, and converting UTC times into the
user's timezone only for the display in the time machine banner).