fix: Correctly handle timezones with negative offset
Because we need to correctly zero-pad the offset's hours, we need to
print the "-" and the actual numbers separately. For this, we check if
the offset is less than zero, and if so, we construct the string ("-" +
zero padded absolute value of offset). However, we computed the absolute
value _before_ checking if the offset is negative, meaning all negative
offsets were actually considered to be positive.

This commit fixes the above by correctly only taking the absolute value
when constructing the rfc3339 string.

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>
Reported-by: kyurime <25387744+kyurime@users.noreply.github.com>