fix: passing only ?before= to /ranking gave top 50 in reverse
When adding the secondary `ORDER BY` expression to the player ranking
query, the sorting (`DESC` or `ASC`) was only applied to the `id`
column, not the `rank` column. However, due to how pagination works with
`?before` parameters (which queries the database by grabbing the first
`limit` items in reverse and then reversing the result again in the
pagination code), this caused pagination for the stats viewer to break.

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>
Fixes: aee50b673d5ef1105703946598c789b8c8aab77b