integrate: Cleanup GD integration
We drop the notion of "cache metadata" in favor of a simple call-site
ratelimiter that only allows one request per day per demon. We also drop
the concept of caching getGJLevels responses (as in, caching the list of
level ids they returned). Lastly, we drop support for grabbing demon
information piecewise: We now either go from "name" to "found a demon
and stored it in the database" or we do not store anything in the
database. The next update will happen in 24 hours anyway, so we get to
try again.

This greatly simplifies the code (to the point where most of it is just
the writing out the sql queries for dumping the 26 fields of a Level
object), and also prevent pointercrate from request-storing the GD
servers (as previously, in the time until the `gd` module managed to
store something in the `_meta` tables, it would simply do requests to
the GD servers for each visit of an uncached demon's page).

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>