Implement a pointercrate-example binary target
This example target aims to be a somewhat minimal setup for a
pointercrate clone. It is a rocket server that has full demonlist
functionalities, but not pointercrate's homepage (or API documentation
or demonlist guidelines). The aim is that it will be useful for setting
up a local development environment that allows one to spin up an actual
server to test out changes.

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>