replace pagination_response! macro with function
Introduce a `Pagination` trait that abstracts away the concept of "thing
that has pagination parameters" and have it offer methods for
manipulating these parameters. This is all that is needed to turn the
macro into a generic function parameterized by `P: Pagination`.

For now, the function has the same horrible implementation as the old
macro - we'll abstract away in follow-up commits.

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>