Move `page` function into `Pagination` trait
Every object that implements `Pagination` had a `page` function that
returns the request page of items, following some validation of the
`limit`, `before` and `after` fields. Add this function to the
`Pagination` trait, so that calls to it can be moved into
`pagination_response`. No longer do the validation inside `page`, but
rather inside `pagination_response` as well.

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>