Give `Pagination::page` sane semantics
Ensure that all this weirdness about "returning list of items in
reverse" and "one extra item indicates that a next page exists" does not
leak out of the actual page implementations, so that
`pagination_response` does not have to deal with it.

This commit only introduces a (hopefully temporary) "compat" function
that translates the current behavior of the various `page` functions
into the new `Pagination::page` semantics, and which is called at the
end of each `page` function.

Also require `Debug` on `Pagination` to be able to do useful log
messages in case things go wrong.

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>