test: Add integration tests for `/demons/listed/` endpoint
Add some tests to ensure that pagination parameters in various
configurations result in the expected responses and the expected
response headers.

Ideally, some of this could be unit-tested, but I do not have the time
to figure out a design that neatly untangles the `PgConnection` objects
from the `Pagination` trait.

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>