Separate `Paginate` trait into two
The `PaginationQuery` trait deals with the actual serialization of the
query parameters, which means that `LinkBuilder` can be unit tested
without mocking out the database, or using dummy implementations for
`page` and friends (since it only needs `parameters` and
`with_parameters` from the old trait). The "async part" of the trait is
now a new `Paginatable<Q: PaginationQuery>` trait, which should be
implemented the the type that actually represents the pagination
response (e.g. on the `Demon` type).

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>