Throttle gd requests to at most 1 per minute
Add a throttle ratelimit that causes the server to wait at least 60
seconds between consecutive requests to the GD servers.

The implementation has to employ a trick where we ratelimit hitting the
global throttle ratelimit. This needs to be done because the `governor`
crate does not allow "refunding" tokens to ratelimiters. E.g. if the
"throttle" ratelimit check returns ok, but the "demon_refresh" check
does not, then we will still wait 1 minutes before even attempting the
next demon refresh. This would be fine if refresh requests for demons
were uniformly distributed, but they are not - refresh request rate is
directly correlated with page views, and page views drop off
exponentially with increasing demon position. This means that page views
of the top 5 or so demons will continuously exhaust the `throttle`
ratelimit, leaving no chance for demons lower on the list to get their
refreshes in. Thus, only allow an individual demon to hit the throttle
rate limit every 10 minutes (number chosen pretty randomly, we can
refine if we still see starvation issues).

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>