Have `/ranking` endpoint use new `score` column
Use the cached scores instead of recomputing them on-demand

Also hopefully fix all pagination weirdness of this endpoint once and
for all. The `index` values were not assigned deterministically, because
the `ROW_INDEX` function was only ordered by score, not by score and id
(which resulted in rows with the same scores being assigned indices
non-deterministically).

This also makes the objects returned by `/ranking` be extensions of
`Player` objects.

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>