Update nation/subdivision scores when needed
The triggers for updating these scores are the same as for player
scores, so we can highjack `update_score` for updates of individual
rows, and `recompute_player_scores` (now simply named
`recompute_scores`) for updates of all scores.

The only one that needs to be treated differently is updating a nations
score when a player's nationality changes - here only the old nation's
score needs to be updated, not the player's as well (although we update
the player score later on anyway).

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>