refactor: Cleanup player nationality patching
Merge the 4 different functions for (re)setting nationality/subdivision
with 4 different SQL UPDATE statements into a single one, with one
UPDATE statement. The cost of this is slightly more complexity at
translating the flat `(nation, subdivision)` tuple that we get as part
of the `PATCH` request into a hierachical `struct Nationality`, but we
need this centralization to be able to easily update the score of a
player's _old_ nationality during a PATCH operation.

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>