Cache stats viewer scores (#129)
* Eliminate `FetchedPlayer`

Seemingly used to be needed due to sqlx restrictions, but this is no
longer the case.

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>

* fix: Remove unnecessary casts and anonymous records

When this code was written, sqlx was in its infancy and didn't support
CITEXT yet. Nowadays it does, so we can clean up some code.

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>

* fix: Remove unused `Demon::shift_up` function

It's been producing rustc warnings for 7 years, its time to let go.

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>

* fix: update doc-comment on `Demon::max_position`

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>

* refactor: Cleanup player nationality patching

Merge the 4 different functions for (re)setting nationality/subdivision
with 4 different SQL UPDATE statements into a single one, with one
UPDATE statement. The cost of this is slightly more complexity at
translating the flat `(nation, subdivision)` tuple that we get as part
of the `PATCH` request into a hierachical `struct Nationality`, but we
need this centralization to be able to easily update the score of a
player's _old_ nationality during a PATCH operation.

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>

* Add `score` column to `players` table

This column will cache a player's stats viewer score, meaning we will no
longer compute the entire leaderboard on-demand when getting a request
to `/ranking`.

This column will be part of the `Player` object, meaning scores are now
accessible outside of the `/ranking` endpoint.

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>

* Recompute player scores as appropriate

When updating/adding records/players/demons, recompute `score` entries
in the `players` table that might be affected by the operation.

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>

* Have `/ranking` endpoint use new `score` column

Use the cached scores instead of recomputing them on-demand

Also hopefully fix all pagination weirdness of this endpoint once and
for all. The `index` values were not assigned deterministically, because
the `ROW_INDEX` function was only ordered by score, not by score and id
(which resulted in rows with the same scores being assigned indices
non-deterministically).

This also makes the objects returned by `/ranking` be extensions of
`Player` objects.

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>

* Add `score` columns to `nationality` and `subdivision` tables

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>

* Update nation/subdivision scores when needed

The triggers for updating these scores are the same as for player
scores, so we can highjack `update_score` for updates of individual
rows, and `recompute_player_scores` (now simply named
`recompute_scores`) for updates of all scores.

The only one that needs to be treated differently is updating a nations
score when a player's nationality changes - here only the old nation's
score needs to be updated, not the player's as well (although we update
the player score later on anyway).

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>

* Use cached scores for nation stats viewer

Have nation's stats viewer use the cached nation scores, instead of
recomputing them all on-demand on every request.

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>

* Use cached scores for heatmap

This is what's been the real performance killer: The queries involved in
constructing the heatmap were so insanely inefficient that you could
literally see the website loading first the heatmap, and the the actual
stats viewer, like its 1992 and your dial up modem only supports a
download speed of 286 bytes per second.

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>

* Drop some now-unused SQL functions

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>

* Ensure no player achieves impossible nation/subdivision combination

A foreign key constraint ignores NULL entries, so things like ('US',
NULL) will still be possible.

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>

* fix: Do not preserve subdivision if player changes nationality

If we receive a player PATCH that changes the country, but does not
specify a new subdivision, reset the subdivision to `None`. This would
previously result in violations of the foreign key constraint added in
the previous commit.

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>

* test: Add various test cases around patching natonalities

Regression test for the previous commit + some other cases, since the
code got refactored in this PR.

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>

* test: Turn `player.rs` into `player/mod.rs`

There will be a separate module dedicated to only score related tests,
so prepare for this by adjusting the module structure.

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>

* test: Add various score related tests

So far we cover the most important scenarios:
- Player score updated when records are accepted/rejected
- Demon verification gives score
- Changing your nationality updates the nationaliy's score

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>

---------

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>