fix: set `min-width: 100%` on stats viewer world svg
Ensure that even if the screen is wider than the svg, it will fill the
entire screen width (potentially zooming in the entire thing until it
does so). We have to use `min-width` instead of `width`, as the latter
would force it to zoom out on mobile device (such that the entire world
map shrinks to be visible on a mobile device), but we do not want that
because then the actual stats viewer panel is rendered too low (because
the map is not tall enough).

Fixes #99

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>