fix: Have player score reset to 0 if last record is removed
If a player's last score-giving record is removed (say because it was on
demon #150 which gets pushed to #151), their score needs to be reset to
0. However, in this scenario, the "select sum over all record giving
scores" will simply not return a row for this player, and thus the
UPDATE query that joins against the result of this select will simply
drop the update for this player. Fix this by instead doing a left join
in the update query so that players with no score-giving records are
correctly assigned a score of 0.

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>