fix: do not display frontend error if player not found
When filtering the record manager by a specific player, do not cause a
`TypeError` by calling a non-existing function (`FormInput.setError`).
Instead use the `errorText` property to set the error message of the
player name input field.

The `TypeError` used to be translated into a frontend java script error
message, which is not very nice.

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>