test: convert thumbnail related integration tests to unittests
I don't know why I have absolutely refused to write unit tests in the
past. For a lot of functionality, integration tests are needed, as the
code under test either lives inside a request handler, or spans
interactions between multiple crates (player claims for example). But
for test cases that only concern themselves with for example a single
endpoint, and then only test functionality that lives in, for example,
`pointercrate-demonlist`, we can just use unit tests.

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>