fix: 500 error when using nation stats viewer
pointercrate#151 changed the definition of the `records` table without
updating a SQL function that was declared as `RETURNS TABLE (LIKE
records)`. This meant that the return type of this function changed to
returning something resembling the new definition of the `records`
table, but the query contained therein did not return something that
matches this new definition (it was missing the raw_footage column).

Fix this by explicitly writing out the table that is returned by the
`best_records_in` function.

Fixes #156

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>