fix: Correctly handle timezones like GMT-2.5 in time machine
Timezones where the offset is not a complete hour (such as GMT-2.5) were
not handled correctly when constructing the RFC3339 date for the `when`
cookie used by the time machine. Instead of being converted to `-02:30`,
it'd get turned into `-2.5:30`, which caused the time machine to break
for everyone in such timezones (as the pointercrate backend simply
ignores malformed `when` cookie values).

Reported-by: https://twitter.com/m1chae1op/status/1812658343221002375
Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>