refactor: split `password_hash` field from `AuthenticatedUser`
Replace it with a `AuthenticationMethod` field. For now, this enum only
contains the `Legacy` variant, indicating that the given user is using
the current (but soon to be legacy) account system. Later, other
authentication methods such as OAuth will add their own variants here.

When matching in on `auth_method`, introduce unreachable `_ => ...` arms
that will return sensible default-errors for functions that that only
operate on legacy accounts (e.g. when doing oauth, you won't be able to
change your password).

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>