refactor: hide `register` endpoint behind `legacy_accounts` feature
Once we support oauth accounts, we want to disable legacy accounts.
However, for testing purposes, as well as for people wanting to run
pointercrate clones, it is still desirable to use the legacy account
system. Thus, hide it behind a cargo feature.

Note that this cargo flag only disables the creation of new legacy
accounts. Pre-existing legacy accounts will continue to be able to log
in.

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>