refactor: Move legacy-account related functionality into module
Takes the refactor of #171 a few steps further. Instead of only turning
`AuthenticationMethod` into an enum, turn `AuthenticatedUser` into an
enum. This enforces at the type level that things like
`invalidate_all_tokens` has an implementation specific to the
authentication method (whereas in the old version, it was possible to
just ignore that function and have it fail with `Unauthorized` on
non-legacy accounts).

Each authentication method will then implement its own module, and
update all the `match` statements that match on `AuthenticatedUser`.

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>