Use new player dropdowns in player edit dialogs
Build dialogs that are similar to the demon edit dialog already in use
for the record editor: Give a player selection dropdown to select a
player from instead of the current paginator based crap. The only
difference is that we need a button, since we might want to set the
player to a non-preexisting player (in which case we cannot trigger the
edit via dropdown selection, as no dropdown selection will occur - the
text field will be used a free-text).

While we're at it, make a slight improvement to the form contained in
the edit dialog, and allow specifying what the name of the only form
field will be. That allows us to match the field name to what the
respective PATCH request expects, meaning we do not need to translate
`data.player` to the correct field name.

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>