fix(demonlist): Player::by_name_or_create returning DatabaseError
Player::by_name_or_create would return a `DatabaseError` due to unique
constraint violation if the name of an existing player object was
passed, yet the capitalization mismatched. I have _no_ idea why sqlx
does not correctly determine the need to treat the parameter in the
query as `CITEXT`, but apparently it needs the explicit case here. Ugh.

Also add regression test (and in general just a few tests for this
function).

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>