Use Signed Double-Submit Cookie to prevent CSRF attacks
Replace the old CSRF-protection based on embedded csrf tokens in html
with session-based signed double-submit cookies, as described in
https://cheatsheetseries.owasp.org/cheatsheets/Cross-Site_Request_Forgery_Prevention_Cheat_Sheet.html#signed-double-submit-cookie-recommended

In our scenario, we reuse the access_token cookie as the second
submission of the session id that's stored inside the CSRF token, as we
store the session id inside the access token.

This has the added advantage that we no longer need to explicitly embed
CSRF tokens into every page on which we want authorization to have some
sort of effect (e.g. to enable list mod to submit approved records
through the standard record submitter).

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>