Add a generation ID to user accounts
The generation ID is encoded into access tokens, and a change of
generation id invalidates access tokens. This gives us a more
straight-forward way to invalidate access tokens, instead of the current
hack of "rehashing the password invalidates the key that signed the
access token because the password hash is part of it".

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>