refactor: Redesign `auth` module to get type-level guarantees
This commit rewrites the authentication system, centered on the
`AuthenticatedUser` struct. By encoding which validation a specific
`AuthenticatedUser` has gone through at the type level, we statically
ensure that, for example, a request made with an API token can never be
used to modify the currently logged in pointercrate account.

The motivation behind this was that there should be no endpoints that
_require_ HTTP Basic Authentication, as with google oauth, the
possibility for this will be going away. The only real option we have
there will be to trust the token stored inside the browser. We thus
start differentiating between API tokens and Browser Tokens, with
browser tokens enjoying higher trust, while also not being usable
outside of the browser context. This means

- The token stored inside the `access_token` cookie can be used to do
  administrative actions on the logged in account (e.g. deletion)
- The `access_token` cookie can ONLY be used by also providing a
  X-CSRF-TOKEN matching the access token
- The `access_token` cookie cannot be uses for Bearer Authentication
- The API tokens that POST /login returns cannot be used for
  browser-based authentication.

The only endpoint that doesn't trust browser tokens is the one for
changing your password, however for oauth accounts this functionality
will not be available anyway.

A user can now be authenticated to 4 staged:
- not authenticated at all (no tokens provided)
- authenticated for API requests (Bearer token in Authorization header,
  OR access_token cookie with valid CSRF token header)
- authenticated for non-mutating/GET requests (access_token cookie
  without CSRF token)
- authenticated for administrative account actions (HTTP Basic Auth or
  access_token cookie with valid CSRF token header).

The non-mutating class exists to get a type level differentiation
between "access_token cookie validated, but no csrf token" and
"access_token cookie valid, and csrf token provided", which is needed
because for GET requests made through browser navigation, we will not
have the CSRF token header (and this is fine, as GET requests are
non-mutating and an attacker cannot read the response of a CSRF attack).

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>