fix: correctly reset error message in claim manager
When selecting a claim where the claimed player does not have any
approved records on the list, we display an error message. However, we
never clear this error message ever again, meaning if we click on
another claim right after, we get both an error message, and an embedded
video (if an approved record exists). Improve this state of affairs in
multiple ways:

- Clear the error message when selecting a claim where the claimed
  player does have an approved error message
- Include the ID of the player without records in the error message (so
  that clicking on multiple claims without records in a row will
  nevertheless have some visual effect)
- Clear the video embed when selecting a claim without any approved
  records.

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>