ui: drop tab-selection CSS class
This class semantically was a container of .tab elements, but there was
no functional reason for having this. Furthermore, there are usecases
where the .tab elements are not all children of the same parent element,
in which case needlessly wrapping them individually in .tab-selection
elements is just clutter and a waste of html elements.

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>