fix: z-index, again
Turns out the reason for drawing the background so weirdly was the stats
viewer: The world map was supposed to go behind the squares, while
everything else is supposed to go in front of it.

So we're back to drawing the background onto an actual div element, so
that the stats viewer world map can be layered behind it. In fact, we're
layering _everything_ behind it, but only show the background if the
screen is wide enough for it to no overlay anything. Let's hope that
goes well.

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>