Cleanup Nationality Statsviewer Code (#230)
* demonlist: Eliminate struct MiniDemon

Instead, use MiniDemonWithPlayers everywhere. It's true that a demon
will have at most one publisher or verifier, but having a whole new type
just to avoid allocating a Vec is a bit much.

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>

* demonlist: store MinimalDemon in MiniDemonWithPlayer

Instead of repeating all the fields of a MinimalDemon, just reuse that
struct. This brings the structure of MiniDemonWithPlayers closer to
MinimalRecordD, its equivalent from the GET /players/[id] endpoint.

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>

* demonlist: store MinimalDemon in BestRecord

This also brings it closer to MinimalRecordD, its rough equivalent from
the FullPlayer object.

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>

* demonlist: Flatten `Nationality` object into `NationalityRecord

Again, to make these structs be more similar to FullPlayer

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>

* demonlist(js): Update nation stats viewer to work with new API

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>

* demonlist(js): default to permalink if no link passed to formatDemon

Literally every callsite was either explicitly passing in the permalink,
or passing it in if some other link was null (e.g. no verification
video). So just move this as a default into the function itself.

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>

---------

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>