Implement Google Oauth2 login + Account Linking (#229)
* add oauth2 feature to Cargo.tomls

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>

* core-pages: add support for async scripts

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>

* ui: drop forced left/right margin on form input contents

If these are wanted, they can be applied on the form-input itself, since
all form inputs we are using are laid out top-to-bottom anyway.

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>

* ui: drop margin on left/right/under/over-lined classes

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>

* css: apply styling to p.error elements independent of parent

Sometimes we might want to use these outside of forms.

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>

* user-api: add utility function for cookie placement

Both /login and /register place CSRF and access tokens in cookies.
Deduplicate the code with a helper function (especially because later
on, oauth will _also_ need to place the same cookies).

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>

* user-pages: add "sign in with google" button to login page

Also add a dummy request handler to the backend which for now always
returns 403 UNAUTHORIZED.

We use the javascript callback version (instead of redirect) of the
google oauth flow so that it's easier to display potential error
messages directly on the login page, as well as to avoid the form data
payload that google would give us if we let it directly POST to our
servers (going through javascript allows us to convert to JSON, which is
easier to deal with).

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>

* add google account id to members table

Mark as UNIQUE to prevent any bugs/race conditions from allowing the
same google account for multiple pointercrate accounts.

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>

* Handle members with linked google accounts in auth module

Introduce a new `AuthenticationType` for these, so that password based
login will automatically be rejected.

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>

* Implement "sign in with google" flow

Have the backend validate credentials it receives via /auth/oauth/google
with the certificates that google publishes, decode the associated JWT
to determine the google account id, and then log in the user who has
that google account id linked by placing the required cookies.

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>

* ui: add "Link with google" button to profile page

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>

* refactor(user): call increment_generation_id from set_password

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>

* user: Implement backend part of google accoutn linking

When reaching the /oauth/google endpoint fully authenticated, and the
provided google credentials are valid, link the authenticated
pointercrate account with the google account. Only do so it the google
account is not already linked to another pointercrate account, and if
the pointercrate account is not already linked to a google account (e.g.
re-linking from one google account to a different one is not possible
right now).

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>

* test: validate no legacy login if google account linked

Add an integration test to ensure that if a google account is linked to
a pointercrate account, that pointercrate account becomes impossible to
log in via username/password combo.

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>

---------

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>