fix: immediately make googleOauthCallback global
This function is referred to by the google button HTML, and needs to be
global by the time the google script runs to initialize itself. By only
making it global in initialize(), it seems that sometimes we lost the
race with the google script, and thus the script didnt see a valid
function assigned to data-callback, and disabled the button.

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>