Include statsviewer rank in GET /players/ responses (#258)
* Add a materialized view for caching player ranks

This allow us to add player ranks to the FullPlayer objects, which will
simplify the stats viewer frontend significantly.

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>

* refresh player_ranks view appropriately

It needs to be refreshed whenever someone's score changes.

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>

* add a test for player merging

I thought that deletes on the players table might get messed up because
the materialized view has an index, but seems like this is not the case.
Good to have a simple test for the merging functionality anyway though.

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>

* return player rank in `Player` objects

This will include it in all pagination responses (we already join with
the nationalities table, so surely this can't be that bad), but more
importantly, we return it in the GET for individual players, which
allows us to simplify the stats viewer a bit.

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>

* js: use API response fields for rank/score in individual statsviewer

With both scores and ranks cached these days, the individual stats
viewer can stop parsing them from the html of the selection panel, and
instead just grab them from the API response. This means that now the
individual stats viewer actually supports .selectArbitrary for selecting
players not currently visible in the pagination panel.

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>

---------

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>